Create evenOdd.cpp
Brief stated to write an optional review program utilizing vectors to perform specific actions on a vector of ints. The directions stated to use a for loop to sum the even ints within a given vector and multiply the odd ints. Then you were told to output the results in a statement.