Create ttt.cpp
main() for tic tac toe game. Brief states to create a c++ program that allows two players to play until they place their tokens three in a row (up, down, across, and diagonally). 

I utilized a header file, and a second c++ file to store the functions for abstraction. All three files are needed for the game to execute. All are uploaded here on GH.