Create ttt_functions.cpp
Tic Tac Toe game functions. 

Note - used a vector to store the board elements. Instead of a blank grid, I placed values in each possible positional placement. To test if the game results in scratch (no winner), I utilized a for loop and a counter to track whether each position had been updated to either 'X' or 'O' and if so, should the test for whether there are three in a row result in false, then the game would be considered a scratch.